Restructured supergrid as an interface for dependency inversion
restored /obj folders to gitignore because of how VS treats them.
