+++ Super-tic-tac-toe/Super-tic-tac-toe/Super-tic-tac-toe.csproj
@@ -63,6 +63,8 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
+    <Compile Include="TicTacToeLogic\Events\Events.cs" />
+    <Compile Include="TicTacToeLogic\Interfaces\ITicTacToeSuperGrid.cs" />
     <Compile Include="TicTacToeLogic\TicTacToeException.cs" />
     <Compile Include="TicTacToeLogic\TicTacToeCell.cs" />
     <Compile Include="TicTacToeLogic\TicTacToeGrid.cs" />
